requirements.txt
llama-index
streamlit
rank-bm25
modelscope
langchain-text-splitters
llama-index==0.12.30
langchain==0.3.23
langchain-ollama==0.3.1
magic-pdf==1.3.8
PyMuPDF==1.24.14
faiss-cpu==1.10.0
streamlit==1.44.1
transformers==4.42.4
torch==2.6.0
numpy==1.26.4
ollama==0.4.7
openai==1.75.0
sentence-transformers>=2.2.2